chore: add `Dockerfile` and config necessary to build/run it
The `Dockerfile` relies on `sqlx`'s offline mode in order to build
without database access. This compels us to add a CI step to check that
the `sqlx-data.json` is up-to-date. An `address` config has been added
so that the docker container can bind to `0.0.0.0` rather than the
default `127.0.0.1`. The `Dockerfile` uses `cargo-chef` to ensure
dependency builds are cached.

In a departure from the source material, we're using alpine from the
get-go (10MB image), and using a whitelist-style `.dockerignore`.