ci: add spec.yaml for deployment to DigitalOcean App Platform
Sadly this needs to be manually sync'd with DO for now, using
`doctl apps update <id> --spec spec.yaml`. It would be nice if DO could
sync itself from a file in the repo before deploying...

The spec is pretty simple, and consists of a service that will build and
serve the Dockerfile, and a database.