refactor: add `ConfigBuilder` to avoid stringly-typed config
This rather egregious refactor of the `config` module allows for base
configuration to be specified in a fully-typed way via a new
`ConfigBuilder` (as seen in `tests/api/helpers.rs`). Additionally, now
that `Config` is entirely read internally, most of the getters have been
removed in favour of using `pub(crate)` fields meaning values can be
moved to the `App` on construction, rather than cloning.