ci: add a migration compatibility test
This currently just checks that the tests from `main` will run with the
new migrations. This needs to be the case during a rolling deployment,
when new application instances will have started and performed
migrations, but old instances will still be running. The bulk of the
check only runs if there are changes to the migration files.