fix: allow ignoring missing migrations on start-up
This needs to be allowed for the app to be able to roll back in the
event of an unhealthy deployment. Thankfully the `MigrateError` enum
covers this case explicitly so we can simply match it and log a warning.

By default this will still error, which is likely to be the desired
behaviour when running locally or in tests. The app spec has been
updated to set the override environment variable to turn it on.