fix: remove native-tls from dependencies
`reqwest` defaults to native TLS, which won't currently build in our
docker image. Although we could make this work, it's preferrable to
simply avoid it. This can be achieved by disabling default features for
`reqwest`, and enabling the `rustls-tls` feature.

This also highlights that we should be testing the docker build, but I
don't want to go down that rabbit hole right now.