fix: update address to full `SocketAddr`
This was missed when the config was moved around. To limit the spread of
configuration, the value for `ADDRESS` has been moved into `spec.yaml`.
This is arguably worse for the default Docker case, but a more paranoid
person might say it's not a good idea to assume that even containers
should bind to `0.0.0.0` by default.