ci: add a merge workflow to update the app spec
This brings the DigitalOcean App spec into continuous deployment. The
`spec.yaml` now includes placeholders that must be substituted
externally (it is invalid YAML, otherwise). `scripts/update-app.sh` can
be used to substitute the placeholders from environment variables
(either pre-set, or prompted if run from a TTY) and update the
DigitalOcean App. The `--wait` flag is used to wait for the deployment
to conclude (and hopefully report an error if it fails, but this hasn't
been tested!).