feat: implement basic authentication for `publish_newsletter`
This adds a `users` table, which the `publish_newsletter` route uses to
check if requests are authenticated. Authentication is expected in HTTP
basic format. Passwords are hashed using Argon2, with some care taken to
ensure the authentication routine is 'constant time' (e.g. does not vary
based on any properties of the input, so attackers can gain no
information such as the [non-]existence of users).