refactor: use `serde` for session serialization
This loosens the dependency on redis. In particular, you don't need to
implement anything redis-specific to store something in the session,
just `serde::{Deserialize,Serialize}`.

This would technically be a breaking change of sorts, since it changes
the session format, which would invalidate any current sessions. We're
way before caring about that though.